feat: implement auto-pass feature with SetAutoPass and CancelAutoPass actions
- Introduced auto-pass functionality allowing players to automatically pass priority until the stack is empty or until the end of their turn.
- Updated GameAction type to include SetAutoPass and CancelAutoPass actions.
- Enhanced GameState to track per-player auto-pass flags.
- Modified ActionButton component to handle auto-pass state and integrate with keyboard shortcuts for toggling auto-pass.
- Implemented logic in the engine to automatically pass priority based on the current auto-pass mode.
- Added tests to ensure correct behavior of the new auto-pass feature.

+++ crates/engine/src/parser/oracle_static.rs
@@ -596,7 +596,7 @@ fn parse_counter_minimum(text: &str) -> Option<(u32, &str)> {
     if let Some((n, rest)) = parse_number(text) {
         let rest = rest.trim_start();
         if let Some(rest) = rest.strip_prefix("or more ") {
-            return Some((n as u32, rest));
+            return Some((n, rest));
         }
     }
     None
